Task5

+++ src/com/zel/idea/Task4.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class Task4 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int x = scanner.nextInt();
-        int y = scanner.nextInt();
-        int z = scanner.nextInt();
+        double x = scanner.nextDouble();
+        double y = scanner.nextDouble();
+        double z = scanner.nextDouble();
         double avg = (x + y + z)/3;
         double halfAvg = avg/2;
         if (Math.floor(halfAvg) > 3)
